pyproject.toml
[build-system]
requires = ["setuptools", "wheel"]
build-backend = "setuptools.build_meta"

[project]
name = "discord-guardian"
version = "0.1.0"
description = "Discord moderation and encouragement bot using Gemini and Firestore"
authors = [{name="You"}]
readme = "README.md"
requires-python = ">=3.10"
dependencies = [
  "discord.py>=2.4.0",
  "python-dotenv>=1.0.1",
  "requests>=2.32.3",
  "google-cloud-firestore>=2.16.0",
  "google-auth>=2.35.0",
]

[project.scripts]
discord-guardian = "guardian.main:main"

[tool.setuptools]
package-dir = {"" = "src"}

[tool.setuptools.packages.find]
where = ["src"]
